change backend logic to be outside of main SDK module
